Having all of the necessary models correctly moved over
