+++ Source/BotDontLie.Common/Models/AzureStorage/GameEntity.cs
