+++ Source/BotDontLie.Common/Models/AzureStorage/PlayerEntity.cs
