+++ Source/BotDontLie.Common/Models/AzureStorage/StatisticsEntity.cs
