+++ Source/BotDontLie.Common/Models/AzureStorage/TeamEntity.cs
