+++ Source/BotDontLie.Common/Models/PlayersResponse.cs
@@ -1,27 +1,27 @@
-﻿// <copyright file="StatsResponse.cs" company="PK Software LLC">
+﻿// <copyright file="PlayersResponse.cs" company="PK Software LLC">
 // Copyright (c) PK Software LLC. All rights reserved.
 // </copyright>
 
-namespace BotDontLie.Models
+namespace BotDontLie.Common.Models
 {
     using System.Collections.Generic;
     using Newtonsoft.Json;
 
     /// <summary>
-    /// This class defines the StatsResponse.
+    /// This class models the players response.
     /// </summary>
-    public class StatsResponse
+    public class PlayersResponse
     {
         /// <summary>
-        /// Gets or sets the necessary statistics.
+        /// Gets or sets the list of players.
         /// </summary>
         [JsonProperty("data")]
 #pragma warning disable CA2227 // Collection properties should be read only
-        public List<Statistic> Statistics { get; set; }
+        public List<Player> Players { get; set; }
 #pragma warning restore CA2227 // Collection properties should be read only
 
         /// <summary>
-        /// Gets or sets the meta data.
+        /// Gets or sets the response metadata.
         /// </summary>
         [JsonProperty("meta")]
         public Meta Meta { get; set; }
