+++ Source/BotDontLie.Common/Models/TeamsAdaptiveSubmitActionData.cs
@@ -2,7 +2,7 @@
 // Copyright (c) PK Software LLC. All rights reserved.
 // </copyright>
 
-namespace BotDontLie.Models
+namespace BotDontLie.Common.Models
 {
     using Microsoft.Bot.Schema;
     using Newtonsoft.Json;
