+++ Source/BotDontLie.Common/Models/TeamsResponse.cs
@@ -1,27 +1,27 @@
-﻿// <copyright file="PlayersResponse.cs" company="PK Software LLC">
+﻿// <copyright file="TeamsResponse.cs" company="PK Software LLC">
 // Copyright (c) PK Software LLC. All rights reserved.
 // </copyright>
 
-namespace BotDontLie.Models
+namespace BotDontLie.Common.Models
 {
     using System.Collections.Generic;
     using Newtonsoft.Json;
 
     /// <summary>
-    /// This class models the players response.
+    /// This class models the Teams API response.
     /// </summary>
-    public class PlayersResponse
+    public class TeamsResponse
     {
         /// <summary>
-        /// Gets or sets the list of players.
+        /// Gets or sets the list of teams.
         /// </summary>
         [JsonProperty("data")]
 #pragma warning disable CA2227 // Collection properties should be read only
-        public List<Player> Players { get; set; }
+        public List<Team> Teams { get; set; }
 #pragma warning restore CA2227 // Collection properties should be read only
 
         /// <summary>
-        /// Gets or sets the meta.
+        /// Gets or sets the meta data.
         /// </summary>
         [JsonProperty("meta")]
         public Meta Meta { get; set; }
